Make proper interface for integrator
I would like to be able to use different ones for testing
